operator>>(Lazy_exac_nt) calls new function read_float_or_quotient
