+++ Number_types/include/CGAL/Lazy_exact_nt.h
@@ -48,6 +48,9 @@
 #include <CGAL/Sqrt_extension_fwd.h>
 #include <CGAL/Kernel/mpl.h>
 
+#include <CGAL/Io/io.h>
+
+
 /*
  * This file contains the definition of the number type Lazy_exact_nt<ET>,
  * where ET is an exact number type (must provide the exact operations needed).
@@ -1302,7 +1305,7 @@ std::istream &
 operator>> (std::istream & is, Lazy_exact_nt<ET> & a)
 {
   ET e;
-  is >> e;
+  internal::read_float_or_quotient(is, e);
   if (is)
     a = e;
   return is;
